requirements.txt
blis==0.7.6
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
cymem==2.0.6
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-WTF==1.0.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.6
nltk==3.6.7
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathy==0.6.1
pdfminer.six==20220319
Pillow==9.0.1
preshed==3.0.6
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.2.0/pt_core_news_sm-3.2.0-py3-none-any.whl
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.15
threadpoolctl==3.1.0
tqdm==4.63.0
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.9
visitor==0.1.3
wasabi==0.9.0
Werkzeug==2.0.3
WTForms==3.0.1
